# Add prefetching code in FP8 GEMM mainloop